+++ packages/vue-query/src/__mocks__/useBaseQuery.ts
@@ -1,9 +1,9 @@
 import { vi } from 'vitest'
 import type { Mock } from 'vitest'
 
-const { useBaseQuery: originImpl, unrefQueryArgs: originalParse } =
-  (await vi.importActual('../useBaseQuery')) as any
+const { useBaseQuery: originImpl } = (await vi.importActual(
+  '../useBaseQuery',
+)) as any
 
 export const useBaseQuery: Mock<(...args: Array<any>) => any> =
   vi.fn(originImpl)
-export const unrefQueryArgs = originalParse