+++ packages/svelte-query-persist-client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @tanstack/svelte-query-persist-client
 
+## 6.0.12
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.11
+  - @tanstack/svelte-query@6.0.10
+
+## 6.0.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/query-persist-client-core@5.91.10
+  - @tanstack/svelte-query@6.0.9
+
 ## 6.0.10
 
 ### Patch Changes